chore: convert apply/node files
