+++ src/apply/node/moveNode.ts
@@ -1,12 +1,14 @@
-const { SyncNode } = require('../../model');
-const { getParent } = require('../../path');
+import { MoveNodeOperation } from 'node.interface';
+import { SyncDoc } from 'types';
+import { SyncNode } from '../../model';
+import { getParent } from '../../path';
 
 /**
  * Applies a move node operation to a SyncDoc.
  *
  * moveNode(doc: SyncDoc, op: MoveNodeOperation): SyncDoc
  */
-const moveNode = (doc, op) => {
+const moveNode = (doc: SyncDoc, op: MoveNodeOperation): SyncDoc => {
     const syncDoc = doc.get('document');
     const [from, fromIndex] = getParent(syncDoc, op.path);
     const [to, toIndex] = getParent(syncDoc, op.newPath);
@@ -17,13 +19,17 @@ const moveNode = (doc, op) => {
 
     const fromChildren = SyncNode.getChildren(from);
     const toChildren = SyncNode.getChildren(to);
-    const toMove = fromChildren.get(fromIndex);
-    const toInsert = toMove.clone();
+    if (fromChildren) {
+        const toMove = fromChildren.get(fromIndex);
+        const toInsert = (toMove.clone() as any) as SyncDoc;
 
-    fromChildren.delete(fromIndex);
-    toChildren.insert(Math.min(toIndex, toChildren.length), [toInsert]);
+        fromChildren.delete(fromIndex);
+        if (toChildren) {
+            toChildren.insert(Math.min(toIndex, toChildren.length), [toInsert]);
+        }
+    }
 
     return doc;
 };
 
-module.exports = moveNode;
+export default moveNode;
